Moved functionality to new component
