+++ frontend/src/Views/CaseView.tsx
@@ -2,11 +2,6 @@ import {
     Tabs,
     Typography,
     Input,
-    EdsProvider,
-    Tooltip,
-    Button,
-    Icon,
-    TextField,
 } from "@equinor/eds-core-react"
 import {
     useEffect,
@@ -15,12 +10,12 @@ import {
 } from "react"
 import { useParams } from "react-router-dom"
 import styled from "styled-components"
-import { edit } from "@equinor/eds-icons"
 import { Project } from "../models/Project"
 import { Case } from "../models/Case"
 import { GetProjectService } from "../Services/ProjectService"
 import { GetCaseService } from "../Services/CaseService"
-import { Modal } from "../Components/Modal"
+import CaseAsset from "../Components/CaseAsset"
+import CaseDescription from "../Components/CaseDescription"
 
 const {
     Panels, Panel,
@@ -47,34 +42,12 @@ const Dg4Field = styled.div`
     display: flex;
 `
 
-const Wrapper = styled.div`
-    display: flex;
-    width: 70%;
-    flex-direction: row;
-`
-
-const ActionsContainer = styled.div`
-    > *:not(:last-child) {
-        margin-right: 0.5rem;
-    }
-`
-
-const CreateAssetForm = styled.form`
-    width: 30rem;
-
-    > * {
-        margin-bottom: 1.5rem;
-    }
-`
-
 function CaseView() {
     const [project, setProject] = useState<Project>()
     const [caseItem, setCase] = useState<Case>()
     const [activeTab, setActiveTab] = useState<number>(0)
     const params = useParams()
     const [dg4DateRec, setDg4DateRec] = useState<Record<string, any>>({})
-    const [caseDescriptionModalIsOpen, setCaseDescriptionModalIsOpen] = useState<boolean>(false)
-    const [caseDescription, setCaseDescription] = useState<string>("")
 
     useEffect(() => {
         (async () => {
@@ -89,23 +62,8 @@ function CaseView() {
         })()
     }, [params.projectId, params.caseId])
 
-    const toggleEditCaseDescriptionModal = () => setCaseDescriptionModalIsOpen(!caseDescriptionModalIsOpen)
-
     const handleTabChange = (index: number) => {
         setActiveTab(index)
-    }
-
-    const handleDescriptionChange: ChangeEventHandler<HTMLInputElement> = async (e) => {
-        setCaseDescription(e.target.value)
-    }
-
-    const submitUpdateDescription = async () => {
-        const caseDto = Case.Copy(caseItem!)
-        caseDto.description = caseDescription
-        const newProject = await GetCaseService().updateCase(caseDto)
-        setProject(newProject)
-        const caseResult = newProject.cases.find((o) => o.id === params.caseId)
-        setCase(caseResult)
     }
 
     const handleDg4FieldChange: ChangeEventHandler<HTMLInputElement> = async (e) => {
@@ -142,29 +100,11 @@ function CaseView() {
             <Tabs activeTab={activeTab} onChange={handleTabChange}>
                 <Panels>
                     <Panel>
-                        <Wrapper>
-                            <Typography
-                                variant="h4"
-                                id="textfield-description"
-                                defaultValue={caseItem?.description}
-                                key={caseItem?.description}
-                            >
-                                {caseItem?.description}
-                                <EdsProvider density="compact">
-                                    <ActionsContainer>
-                                        <Tooltip title="Edit description">
-                                            <Button
-                                                variant="ghost_icon"
-                                                aria-label={`Edit ${caseItem?.description}`}
-                                                onClick={toggleEditCaseDescriptionModal}
-                                            >
-                                                <Icon data={edit} />
-                                            </Button>
-                                        </Tooltip>
-                                    </ActionsContainer>
-                                </EdsProvider>
-                            </Typography>
-                        </Wrapper>
+                        <CaseDescription
+                            caseItem={caseItem}
+                            setProject={setProject}
+                            setCase={setCase}
+                        />
                     </Panel>
                 </Panels>
                 <Panels>
@@ -182,35 +122,14 @@ function CaseView() {
                         </Dg4Field>
                     </Panel>
                 </Panels>
+                <CaseAsset
+                    caseItem={caseItem}
+                    project={project}
+                    setProject={setProject}
+                    setCase={setCase}
+                    caseId={params.caseId}
+                />
             </Tabs>
-            <Modal isOpen={caseDescriptionModalIsOpen} title="Edit description" shards={[]}>
-                <CreateAssetForm>
-                    <TextField
-                        label="Description"
-                        id="description"
-                        name="description"
-                        placeholder={caseItem?.description}
-                        onChange={handleDescriptionChange}
-                    />
-                    <div>
-                        <Button
-                            type="submit"
-                            onClick={submitUpdateDescription}
-                            disabled={caseDescription === ""}
-                        >
-                            Submit
-                        </Button>
-                        <Button
-                            type="button"
-                            color="secondary"
-                            variant="ghost"
-                            onClick={toggleEditCaseDescriptionModal}
-                        >
-                            Cancel
-                        </Button>
-                    </div>
-                </CreateAssetForm>
-            </Modal>
         </CaseViewDiv>
     )
 }
